fix(margin): align market data contracts
Audit all current Margin Market Data routes, sign USER_DATA queries, remove stale parameters, validate filters and receive windows, preserve undefined-unit timestamps as int64, correct public names, and add deterministic contract coverage.

+++ Binance.Api/Margin/Responses/BinanceIsolatedMarginTier.cs
@@ -13,7 +13,7 @@ public record BinanceIsolatedMarginTier
     /// <summary>
     /// Tier
     /// </summary>
-    public int Tier { get; set; }
+    public long Tier { get; set; }
 
     /// <summary>
     /// Effective multiple
